# openapi.yaml
openapi: 3.0.3
info:
  title: Go Email Template Sender API
  version: 1.0.0
  description: A dynamic templating, styling, data-binding, and email delivery service built with Hexagonal Architecture.
servers:
  - url: http://localhost:8080
    description: Local development server

tags:
  - name: Contexts
    description: Arbitrary JSON data context store
  - name: Stylesheets
    description: CSS stylesheet management
  - name: Templates
    description: Dynamic HTML document templates
  - name: Bindings
    description: Data source query bindings
  - name: Deliveries
    description: Email delivery and dispatch orchestrator

paths:
  # ==========================================
  # Contexts
  # ==========================================
  /contexts:
    post:
      tags:
        - Contexts
      summary: Create a new Context
      operationId: createContext
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateContextRequest'
      responses:
        '201':
          description: Context created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdResponse'
    get:
      tags:
        - Contexts
      summary: List Contexts
      operationId: listContexts
      parameters:
        - name: customer_name
          in: query
          description: Filter by customer name
          required: false
          schema:
            type: string
      responses:
        '200':
          description: A list of Contexts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ContextResponse'

  /contexts/{id}:
    get:
      tags:
        - Contexts
      summary: Get a Context by ID
      operationId: getContext
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Context found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContextResponse'
        '404':
          description: Context not found
    put:
      tags:
        - Contexts
      summary: Fully replace an existing Context
      operationId: updateContext
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateContextRequest'
      responses:
        '204':
          description: Context updated successfully
        '404':
          description: Context not found

  # ==========================================
  # Stylesheets
  # ==========================================
  /stylesheets:
    post:
      tags:
        - Stylesheets
      summary: Create a new Stylesheet
      operationId: createStylesheet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStylesheetRequest'
      responses:
        '201':
          description: Stylesheet created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdResponse'
    get:
      tags:
        - Stylesheets
      summary: List Stylesheets
      operationId: listStylesheets
      parameters:
        - name: name
          in: query
          description: Filter by stylesheet name
          required: false
          schema:
            type: string
      responses:
        '200':
          description: A list of Stylesheets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StylesheetResponse'

  /stylesheets/{id}:
    get:
      tags:
        - Stylesheets
      summary: Get Stylesheet by ID
      operationId: getStylesheet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Stylesheet found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StylesheetResponse'
        '404':
          description: Stylesheet not found
    put:
      tags:
        - Stylesheets
      summary: Fully replace an existing Stylesheet
      operationId: updateStylesheet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStylesheetRequest'
      responses:
        '204':
          description: Stylesheet updated successfully
        '404':
          description: Stylesheet not found

  # ==========================================
  # Templates
  # ==========================================
  /templates:
    post:
      tags:
        - Templates
      summary: Create a new Template
      operationId: createTemplate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTemplateRequest'
      responses:
        '201':
          description: Template created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdResponse'
    get:
      tags:
        - Templates
      summary: List Templates
      operationId: listTemplates
      parameters:
        - name: name
          in: query
          description: Filter by template name
          required: false
          schema:
            type: string
      responses:
        '200':
          description: A list of Templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TemplateResponse'

  /templates/{id}:
    get:
      tags:
        - Templates
      summary: Get Template by ID
      operationId: getTemplate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Template found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateResponse'
        '404':
          description: Template not found
    put:
      tags:
        - Templates
      summary: Update Template document content or stylesheet
      operationId: updateTemplate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTemplateRequest'
      responses:
        '204':
          description: Template updated successfully
        '404':
          description: Template not found

  /templates/{id}/render:
    get:
      tags:
        - Templates
      summary: Render a Template with linked CSS injected
      operationId: renderTemplate
      description: Returns the raw rendered HTML document after safely injecting linked stylesheet CSS into the head section.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Rendered HTML document
          content:
            text/html:
              schema:
                type: string
        '404':
          description: Template or linked stylesheet not found

  # ==========================================
  # Bindings
  # ==========================================
  /bindings:
    post:
      tags:
        - Bindings
      summary: Create a new Binding
      operationId: createBinding
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBindingRequest'
      responses:
        '201':
          description: Binding created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdResponse'
    get:
      tags:
        - Bindings
      summary: List Bindings
      operationId: listBindings
      parameters:
        - name: name
          in: query
          description: Filter by binding name
          required: false
          schema:
            type: string
      responses:
        '200':
          description: A list of Bindings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BindingResponse'

  /bindings/{id}:
    get:
      tags:
        - Bindings
      summary: Get Binding by ID
      operationId: getBinding
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Binding found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BindingResponse'
        '404':
          description: Binding not found
    put:
      tags:
        - Bindings
      summary: Fully replace an existing Binding
      operationId: updateBinding
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBindingRequest'
      responses:
        '204':
          description: Binding updated successfully
        '404':
          description: Binding not found

  # ==========================================
  # Deliveries
  # ==========================================
  /deliveries:
    post:
      tags:
        - Deliveries
      summary: Trigger email delivery dispatch
      operationId: createDelivery
      description: Resolves contexts using the binding query, renders HTML template with Handlebars and CSS, and dispatches emails via SMTP.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDeliveryRequest'
      responses:
        '202':
          description: Delivery dispatch accepted and started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeliveryResponse'
        '400':
          description: Invalid request parameters
        '500':
          description: Delivery dispatch failure

# ==========================================
# Reusable Schemas
# ==========================================
components:
  schemas:
    IdResponse:
      type: object
      required:
        - id
      properties:
        id:
          type: string
          format: uuid
          example: "6664076e-7253-4dd6-b768-3bdd81b699d1"

    CreateContextRequest:
      type: object
      required:
        - reference_id
        - customer_name
        - payload
        - email_jsonpath
      properties:
        reference_id:
          type: string
          example: "REF-1001"
        customer_name:
          type: string
          example: "Acme Corp"
        payload:
          type: string
          description: Arbitrary JSON string payload
          example: "{\"user\": {\"name\": \"Alice\", \"email\": \"alice@acme.com\"}}"
        email_jsonpath:
          type: string
          description: JSONPath expression to extract recipient email address from payload
          example: "$.user.email"

    ContextResponse:
      type: object
      required:
        - id
        - reference_id
        - customer_name
        - payload
        - email_jsonpath
      properties:
        id:
          type: string
          format: uuid
        reference_id:
          type: string
        customer_name:
          type: string
        payload:
          type: string
        email_jsonpath:
          type: string

    CreateStylesheetRequest:
      type: object
      required:
        - name
        - code
        - css_content
      properties:
        name:
          type: string
          example: "Corporate Blue Theme"
        code:
          type: string
          example: "STYLE_CORP_01"
        css_content:
          type: string
          example: "h1 { color: #003366; font-family: sans-serif; }"

    StylesheetResponse:
      type: object
      required:
        - id
        - name
        - code
        - css_content
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        code:
          type: string
        css_content:
          type: string

    CreateTemplateRequest:
      type: object
      required:
        - name
        - code
        - html_content
        - subject
      properties:
        name:
          type: string
          example: "Monthly Invoice Template"
        code:
          type: string
          example: "TMPL_INV_01"
        html_content:
          type: string
          description: Raw HTML content with Handlebars placeholders
          example: "<h1>Invoice for {{user.name}}</h1>"
        subject:
          type: string
          description: Email subject line
          example: "Your Monthly Invoice"
        stylesheet_id:
          type: string
          format: uuid
          nullable: true
          description: Optional foreign key linked to a Stylesheet

    TemplateResponse:
      type: object
      required:
        - id
        - name
        - code
        - version
        - html_content
        - subject
        - stylesheet_id
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        code:
          type: string
        version:
          type: integer
          example: 1
        html_content:
          type: string
        subject:
          type: string
        stylesheet_id:
          type: string
          format: uuid
          nullable: true

    CreateBindingRequest:
      type: object
      required:
        - name
        - query
        - template_id
      properties:
        name:
          type: string
          example: "Active Customer Invoices Binding"
        query:
          type: string
          description: SQL query statement selecting contexts from contexts table
          example: "SELECT id, reference_id, customer_name, payload::text, email_jsonpath FROM contexts WHERE payload->'user'->>'status' = 'ACTIVE'"
        template_id:
          type: string
          format: uuid

    BindingResponse:
      type: object
      required:
        - id
        - name
        - query
        - template_id
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        query:
          type: string
        template_id:
          type: string
          format: uuid

    CreateDeliveryRequest:
      type: object
      required:
        - template_id
        - binding_id
      properties:
        template_id:
          type: string
          format: uuid
        binding_id:
          type: string
          format: uuid

    DeliveryResponse:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          example: "dispatched"